+++ src/config/gen.ts
@@ -1,5 +1,7 @@
 import * as T from "../types";
 import { downloadFile, trace } from "../util";
+import { Mutex } from 'async-mutex'
+import env from '../env';
 import logger from '../logger';
 import { saneLoggerConfig, quietLoggerConfig } from './logger';
 const TOML = require('@iarna/toml')
@@ -18,7 +20,7 @@ const mkdirIdempotent = dir => fs.access(dir).catch(() => {
   fs.mkdir(dir, { recursive: true })
 })
 
-const tempDirBase = path.join(process.env.TRYORAMA_STORAGE || os.tmpdir(), 'try-o-rama/')
+const tempDirBase = path.join(env.tempStorage || os.tmpdir(), 'try-o-rama/')
 mkdirIdempotent(tempDirBase)
 
 const tempDir = async () => {
@@ -44,6 +46,8 @@ export const dna = (location, id?, opts = {}): T.DnaConfig => {
   return { file: location, id, ...opts }
 }
 
+const downloadMutex = new Mutex()
+
 /**
  * 1. If a dna config object contains a URL in the path, download the file to a temp directory, 
  *     and rewrite the path to point to downloaded file.
@@ -57,8 +61,13 @@ export const resolveDna = async (inputDna: T.DnaConfig, providedUuid: string): P
   }
   if (dna.file.match(/^https?:/)) {
     const dnaPath = path.join(await dnaDir(), dna.id + '.dna.json')
-    await downloadFile({ url: dna.file, path: dnaPath, overwrite: false })
-    dna.file = dnaPath
+    const release = await downloadMutex.acquire()
+    try {
+      await downloadFile({ url: dna.file, path: dnaPath, overwrite: false })
+    } finally {
+      dna.file = dnaPath
+      release()
+    }
   }
 
   dna.id = dna.uuid ? `${dna.id}::${dna.uuid}` : dna.id
@@ -204,8 +213,7 @@ export const genInstanceConfig = async ({ instances }, { configDir, adminPort, z
       agent: instance.agent.id,
       dna: resolvedDna.id,
       storage: {
-        type: 'file',
-        path: path.join(configDir, instance.id)
+        type: 'memory'
       }
     })
     zomeInterface.instances.push({ id: instance.id })